+++ common/CompilerDetect.h
@@ -37,8 +37,8 @@
 #define MPT_CLANG_AT_LEAST(major,minor,patch)        (MPT_COMPILER_CLANG_VERSION >= MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 #define MPT_CLANG_BEFORE(major,minor,patch)          (MPT_COMPILER_CLANG_VERSION <  MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 
-#if MPT_CLANG_BEFORE(3,4,0)
-#error "clang version 3.4 required"
+#if MPT_CLANG_BEFORE(3,6,0)
+#error "clang version 3.6 required"
 #endif
 
 #if defined(__clang_analyzer__)