+++ common/misc_util.h
@@ -256,7 +256,7 @@ template <class T>
 inline void MemsetZero(T &a)
 {
 	static_assert(std::is_pointer<T>::value == false, "Won't memset pointers.");
-#if MPT_GCC_BEFORE(5,1,0) || MPT_CLANG_BEFORE(3,5,0) || (MPT_COMPILER_CLANG && defined(__GLIBCXX__))
+#if MPT_GCC_BEFORE(5,1,0) || (MPT_COMPILER_CLANG && defined(__GLIBCXX__))
 	MPT_STATIC_ASSERT(std::is_standard_layout<T>::value);
 	MPT_STATIC_ASSERT(std::is_trivial<T>::value); // approximation
 #else // default
@@ -277,7 +277,7 @@ template <typename Tdst, typename Tsrc>
 MPT_FORCEINLINE Tdst bit_cast(const Tsrc & src) noexcept
 {
 	MPT_STATIC_ASSERT(sizeof(Tdst) == sizeof(Tsrc));
-#if MPT_GCC_BEFORE(5,1,0) || MPT_CLANG_BEFORE(3,5,0) || (MPT_COMPILER_CLANG && defined(__GLIBCXX__))
+#if MPT_GCC_BEFORE(5,1,0) || (MPT_COMPILER_CLANG && defined(__GLIBCXX__))
 	MPT_STATIC_ASSERT(std::is_trivial<Tdst>::value); // approximation
 	MPT_STATIC_ASSERT(std::is_trivial<Tsrc>::value); // approximation
 #else // default